+++ r-package/tests/tests_rafa/test_rafa.R
@@ -253,7 +253,6 @@ Sys.setenv(NOT_CRAN = "true")
 
 
 
-function_coverage(fun='check_connection', test_file("tests/testthat/test-check_connection.R"))
 function_coverage(fun='download_metadata', test_file("tests/testthat/test-download_metadata.R"))
 function_coverage(fun='list_geobr', test_file("tests/testthat/test-list_geobr.R"))
 function_coverage(fun='lookup_muni', test_file("tests/testthat/test-lookup_muni.R"))